AG-11128 Refactor channel scenes to extend shared abstract class
