+++ packages/ag-charts-enterprise/src/features/annotations/scenes/channelScene.ts
@@ -1,0 +1,92 @@
+import { _Scene } from 'ag-charts-community';
+
+import type { LineCoords } from '../annotationTypes';
+import { Annotation } from './annotation';
+import type { Handle } from './handle';
+import { CollidableLine } from './shapes';
+
+export abstract class Channel<
+    Datum extends { background: { fill?: string; fillOpacity?: number }; locked?: boolean; visible?: boolean },
+> extends Annotation {
+    protected handles: { [key: string]: Handle } = {};
+    protected seriesRect?: _Scene.BBox;
+
+    protected topLine = new CollidableLine();
+    protected bottomLine = new CollidableLine();
+    protected background = new _Scene.Path({ zIndex: -1 });
+
+    public update(datum: Datum, seriesRect: _Scene.BBox, top?: LineCoords, bottom?: LineCoords) {
+        const { locked, visible } = datum;
+
+        this.locked = locked ?? false;
+        this.seriesRect = seriesRect;
+
+        if (top == null || bottom == null) {
+            this.visible = false;
+            return;
+        } else {
+            this.visible = visible ?? true;
+        }
+
+        this.updateLines(datum, top, bottom);
+        this.updateHandles(datum, top, bottom);
+        this.updateBackground(datum, top, bottom);
+    }
+
+    override toggleActive(active: boolean) {
+        this.toggleHandles(active);
+        for (const node of Object.values(this.handles)) {
+            node.toggleActive(active);
+        }
+    }
+
+    override stopDragging() {
+        const { activeHandle, handles } = this;
+        if (activeHandle == null) return;
+
+        handles[activeHandle].toggleDragging(false);
+    }
+
+    override getCursor() {
+        if (this.activeHandle == null) return 'pointer';
+        return this.handles[this.activeHandle].getCursor();
+    }
+
+    override containsPoint(x: number, y: number) {
+        const { handles, seriesRect, topLine, bottomLine } = this;
+
+        this.activeHandle = undefined;
+
+        for (const [handle, child] of Object.entries(handles)) {
+            if (child.containsPoint(x, y)) {
+                this.activeHandle = handle;
+                return true;
+            }
+        }
+
+        x -= seriesRect?.x ?? 0;
+        y -= seriesRect?.y ?? 0;
+
+        return topLine.containsPoint(x, y) || bottomLine.containsPoint(x, y);
+    }
+
+    protected abstract updateLines(datum: Datum, top: LineCoords, bottom: LineCoords): void;
+
+    protected abstract updateHandles(datum: Datum, top: LineCoords, bottom: LineCoords): void;
+
+    protected updateBackground(datum: Datum, top: LineCoords, bottom: LineCoords) {
+        const { background } = this;
+
+        background.path.clear();
+        background.path.moveTo(top.x1, top.y1);
+        background.path.lineTo(top.x2, top.y2);
+        background.path.lineTo(bottom.x2, bottom.y2);
+        background.path.lineTo(bottom.x1, bottom.y1);
+        background.path.closePath();
+        background.checkPathDirty();
+        background.setProperties({
+            fill: datum.background.fill,
+            fillOpacity: datum.background.fillOpacity,
+        });
+    }
+}
